refactor: change catalogo filters logic

+++ src/main/java/com/ufcg/psoft/commerce/service/cafe/CafeServiceImpl.java
@@ -4,6 +4,7 @@ import com.ufcg.psoft.commerce.dto.cliente.ClientePostPutRequestDTO;
 import com.ufcg.psoft.commerce.exception.CafeNaoExisteException;
 import com.ufcg.psoft.commerce.exception.ClienteNaoExisteException;
 import com.ufcg.psoft.commerce.exception.CodigoDeAcessoInvalidoException;
+import com.ufcg.psoft.commerce.exception.FornecedorNaoForneceCafeException;
 import com.ufcg.psoft.commerce.exception.InteresseEmCafeDisponivelException;
 import com.ufcg.psoft.commerce.model.Cliente;
 import com.ufcg.psoft.commerce.service.fornecedor.FornecedorService;
@@ -21,6 +22,8 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -49,8 +52,11 @@ public class CafeServiceImpl implements CafeService{
     @Override
     public void remover(Long idCafe, Long idFornecedor, String codigoAcesso) {
         fornecedorService.verificaFornecedor(idFornecedor, codigoAcesso);
-
+        
         Cafe cafe = cafeRepository.findById(idCafe).orElseThrow(CafeNaoExisteException::new);
+        if(cafe.getFornecedor().getId() != idFornecedor){
+            throw new FornecedorNaoForneceCafeException();
+        };
         cafeRepository.delete(cafe);
     }
 
@@ -70,6 +76,9 @@ public class CafeServiceImpl implements CafeService{
         fornecedorService.verificaFornecedor(idFornecedor, codigoAcesso);
 
         Cafe cafe = cafeRepository.findById(idCafe).orElseThrow(CafeNaoExisteException::new);
+        if(cafe.getFornecedor().getId() != idFornecedor){
+            throw new FornecedorNaoForneceCafeException();
+        };
         modelMapper.map(cafePostPutRequestDTO, cafe);
         cafeRepository.save(cafe);
         return modelMapper.map(cafe, CafeResponseDTO.class);
@@ -81,6 +90,16 @@ public class CafeServiceImpl implements CafeService{
         return cafes.stream()
                 .map(CafeResponseDTO::new)
                 .collect(Collectors.toList());
+    }
+
+    private List<Cafe> ordenaPorDisponibilidade(List<Cafe> cafes){
+        Collections.sort(cafes, new Comparator<Cafe>() {
+            @Override
+            public int compare(Cafe c1, Cafe c2) {
+                return Boolean.compare(c2.isDisponivel(), c1.isDisponivel());
+            }
+        });
+        return cafes;
     }
 
     @Override
@@ -99,18 +118,14 @@ public class CafeServiceImpl implements CafeService{
         List<Cafe> cafes;
 
         if(cliente.getAssinatura().equals(TipoAssinatura.PREMIUM)) {
-            cafes = cafeRepository.findByDisponivel(true);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivel(false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findAll();
+
         } else {
-            cafes = cafeRepository.findByQualidadeAndDisponivel(QualidadeCafe.NORMAL, true);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivel(QualidadeCafe.NORMAL,false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidade(QualidadeCafe.NORMAL);
+
         }
+
+        cafes = ordenaPorDisponibilidade(cafes);
 
 
         return cafes.stream()
@@ -140,54 +155,23 @@ public class CafeServiceImpl implements CafeService{
 
         if(tipo == null && origem.isBlank() && perfil.isBlank()) {
             cafes = cafeRepository.findByDisponivel(true);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivel(false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
         } else if(origem.isBlank() && perfil.isBlank()) {
-            cafes = cafeRepository.findByTipoAndDisponivel(tipo, true);
-            List<Cafe> indisponiveis = cafeRepository.findByTipoAndDisponivel(tipo,false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByTipo(tipo);
         } else if(tipo == null && origem.isBlank()) {
-            cafes = cafeRepository.findByDisponivelAndPerfilContainingIgnoreCase(true, perfil);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndPerfilContainingIgnoreCase(false, perfil);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByPerfilContainingIgnoreCase(perfil);
         } else if(tipo == null && perfil.isBlank()) {
-            cafes = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCase(true, origem);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCase(false, origem);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByOrigemContainingIgnoreCase(origem);
         } else if(tipo == null) {
-            cafes = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(true, origem, perfil);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(false, origem, perfil);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
-            
+            cafes = cafeRepository.findByOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(origem, perfil);
         } else if(origem.isBlank()) {
-            cafes = cafeRepository.findByDisponivelAndPerfilContainingIgnoreCaseAndTipo(true, perfil, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndPerfilContainingIgnoreCaseAndTipo(false, perfil, tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByPerfilContainingIgnoreCaseAndTipo(perfil, tipo);
         } else if(perfil.isBlank()){
-            cafes = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndTipo(true, origem, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndTipo(false, origem, tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByOrigemContainingIgnoreCaseAndTipo(origem, tipo);
         } else {
-            cafes = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(true, origem, perfil, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(false, origem, perfil, tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(origem, perfil, tipo);
         }
+
+        cafes = ordenaPorDisponibilidade(cafes);
 
         return cafes.stream()
                 .map(CafeResponseDTO::new)
@@ -198,56 +182,24 @@ public class CafeServiceImpl implements CafeService{
         List<Cafe> cafes;
 
         if(tipo == null && origem.isBlank() && perfil.isBlank()) {
-            cafes = cafeRepository.findByQualidadeAndDisponivel(QualidadeCafe.NORMAL, true);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivel(QualidadeCafe.NORMAL,false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
-            
+            cafes = cafeRepository.findByQualidade(QualidadeCafe.NORMAL);
         } else if(origem.isBlank() && perfil.isBlank()) {
-            cafes = cafeRepository.findByQualidadeAndTipoAndDisponivel(QualidadeCafe.NORMAL, tipo, true);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndTipoAndDisponivel(QualidadeCafe.NORMAL,tipo,false);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndTipo(QualidadeCafe.NORMAL, tipo);
         } else if(tipo == null && origem.isBlank()) {
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL, true, perfil);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL,false,perfil);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL, perfil);
         } else if(tipo == null && perfil.isBlank()) {
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCase(QualidadeCafe.NORMAL ,true, origem);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCase(QualidadeCafe.NORMAL,false,origem);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndOrigemContainingIgnoreCase(QualidadeCafe.NORMAL,  origem);
         } else if(tipo == null) {
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL, true, origem, perfil);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL,false,origem,perfil);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCase(QualidadeCafe.NORMAL, origem, perfil);
         } else if(origem.isBlank()) {
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, true, perfil, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL,false,perfil,tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, perfil, tipo);
         } else if(perfil.isBlank()){
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, true, origem, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL,false,origem,tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndOrigemContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, origem, tipo);
         } else {
-            cafes = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, true, origem, perfil, tipo);
-            List<Cafe> indisponiveis = cafeRepository.findByQualidadeAndDisponivelAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL,false,origem,perfil,tipo);
-            for (Cafe cafe : indisponiveis) {
-                cafes.add(cafe);
-            }
+            cafes = cafeRepository.findByQualidadeAndOrigemContainingIgnoreCaseAndPerfilContainingIgnoreCaseAndTipo(QualidadeCafe.NORMAL, origem, perfil, tipo);
         }
 
+        cafes = ordenaPorDisponibilidade(cafes);
 
         return cafes.stream()
                 .map(CafeResponseDTO::new)
@@ -285,6 +237,9 @@ public class CafeServiceImpl implements CafeService{
     public CafeResponseDTO alterarDisponibilidadeCafe(Long idCafe, Long idFornecedor, String codigoAcesso, boolean disponibilidade){
         fornecedorService.verificaFornecedor(idFornecedor, codigoAcesso);
         Cafe cafe = recuperaCafe(idCafe);
+        if(cafe.getFornecedor().getId() != idFornecedor){
+            throw new FornecedorNaoForneceCafeException();
+        }
         if(disponibilidade && !cafe.isDisponivel()){
             notificaClientesInteressados(cafe);
         }
